# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AdmitGraph API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/research/plan": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Create Plan",
        "operationId": "create_plan_api_v1_research_plan_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchPlanCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchPlanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/runs": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Create Run",
        "operationId": "create_run_api_v1_research_runs_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchRunCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchPlanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/research/runs/{run_id}": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_v1_research_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchPlanDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/runs/{run_id}/events": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Run Events",
        "operationId": "run_events_api_v1_research_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchEventsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Cancel Run",
        "operationId": "cancel_run_api_v1_research_runs__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Cancel Run Api V1 Research Runs  Run Id  Cancel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResearchEventsOut": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/ResearchStepOut"
            },
            "type": "array",
            "title": "Steps"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "steps"
        ],
        "title": "ResearchEventsOut"
      },
      "ResearchPlanCreate": {
        "properties": {
          "intake_year": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2100.0,
                "minimum": 2020.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Intake Year"
          },
          "goal": {
            "additionalProperties": true,
            "type": "object",
            "title": "Goal",
            "default": {}
          }
        },
        "type": "object",
        "title": "ResearchPlanCreate"
      },
      "ResearchPlanDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "planned_queries": {
            "items": {},
            "type": "array",
            "title": "Planned Queries"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "mode",
          "planned_queries",
          "error_message",
          "created_at",
          "started_at",
          "completed_at"
        ],
        "title": "ResearchPlanDetail"
      },
      "ResearchPlanOut": {
        "properties": {
          "research_plan_id": {
            "type": "string",
            "format": "uuid",
            "title": "Research Plan Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "research_plan_id",
          "status"
        ],
        "title": "ResearchPlanOut"
      },
      "ResearchRunCreate": {
        "properties": {
          "intake_year": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2100.0,
                "minimum": 2020.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Intake Year"
          },
          "goal": {
            "additionalProperties": true,
            "type": "object",
            "title": "Goal",
            "default": {}
          }
        },
        "type": "object",
        "title": "ResearchRunCreate"
      },
      "ResearchStepOut": {
        "properties": {
          "step_key": {
            "type": "string",
            "title": "Step Key"
          },
          "service_name": {
            "type": "string",
            "title": "Service Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "output": {
            "additionalProperties": true,
            "type": "object",
            "title": "Output"
          }
        },
        "type": "object",
        "required": [
          "step_key",
          "service_name",
          "status",
          "error_message",
          "output"
        ],
        "title": "ResearchStepOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}